+++ geotrek/core/templates/core/sql/post_90_defaults.sql
@@ -36,6 +36,7 @@ ALTER TABLE core_topology ALTER COLUMN "offset" SET DEFAULT 0.0;
 ALTER TABLE core_topology ALTER COLUMN kind SET DEFAULT '';
 ALTER TABLE core_topology ALTER COLUMN "length" SET DEFAULT 0.0;
 ALTER TABLE core_topology ALTER COLUMN geom_need_update SET DEFAULT FALSE;
+ALTER TABLE core_topology ALTER COLUMN decoupled SET DEFAULT FALSE;
 -- geom
 ALTER TABLE core_topology ALTER COLUMN uuid SET DEFAULT gen_random_uuid();
 -- geom_3d