+++ geotrek/land/views.py
@@ -77,6 +77,7 @@ class PhysicalEdgeFormatList(MapEntityFormat, PhysicalEdgeList):
         "districts",
         "areas",
         "uuid",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]
 
@@ -146,6 +147,7 @@ class LandEdgeFormatList(MapEntityFormat, LandEdgeList):
         "areas",
         "uuid",
         "length_2d",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]
 
@@ -212,6 +214,7 @@ class CirculationEdgeFormatList(MapEntityFormat, CirculationEdgeList):
         "areas",
         "uuid",
         "length_2d",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]
 
@@ -276,6 +279,7 @@ class CompetenceEdgeFormatList(MapEntityFormat, CompetenceEdgeList):
         "areas",
         "uuid",
         "length_2d",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]
 
@@ -340,6 +344,7 @@ class WorkManagementEdgeFormatList(MapEntityFormat, WorkManagementEdgeList):
         "areas",
         "uuid",
         "length_2d",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]
 
@@ -404,6 +409,7 @@ class SignageManagementEdgeFormatList(MapEntityFormat, SignageManagementEdgeList
         "areas",
         "uuid",
         "length_2d",
+        "decoupled",
         *AltimetryMixin.COLUMNS,
     ]