+++ docs/changelog.rst
@@ -9,6 +9,10 @@ CHANGELOG
 
 - Topologies are no longer dependant on the path network: they can now exist independently or be based on paths, regardless of the dynamic segmentation setting.
 
+**Improvements**
+
+* Add ``decoupled`` boolean field to the Topology model
+
 
 2.123.1         (2026-01-29)
 ----------------------------